apps/nexo-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0.post1
pydantic>=2.5.2
firebase-admin>=7.3.0
python-dotenv>=1.0.0
PyMuPDF>=1.27.2.2
loguru>=0.7.3
slowapi>=0.1.9
requests>=2.32.5
python-multipart>=0.0.22
openai>=2.29.0
reportlab>=4.4.10
python-docx>=1.2.0
stripe>=14.4.1
mercadopago>=2.3.0
psycopg2-binary>=2.9.9